[luci] Separate ConnectNode class
This will separate ConnectNode class to it's own file.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
